# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Auth Service API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/v1/auth",
    "paths": {
        "/tokens": {
            "get": {
                "description": "Генерирует новую пару access и refresh токенов на основе GUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Генерация новых JWT токенов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор (UUID)",
                        "name": "guid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth_service_internal_dto.CreatedTokensResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный GUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tokens/refresh": {
            "post": {
                "description": "Обновляет access и refresh токены, если токены являются валидными.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление JWT токенов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор (UUID)",
                        "name": "guid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Access и Refresh токены",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_service_internal_dto.TokensRefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth_service_internal_dto.CreatedTokensResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или GUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Некорректные токены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth_service_internal_dto.CreatedTokensResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "auth_service_internal_dto.TokensRefreshRequest": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        }
    }
}